+++ packages/dataviews/CHANGELOG.md
@@ -11,9 +11,10 @@
 ### Internal
 
 -   Remove some unused dependencies ([#62010](https://github.com/WordPress/gutenberg/pull/62010)).
+
 ### Enhancement
 
--   Add `getLabel` prop in Actions API to support labels that use information from the selected items. ([#61942](https://github.com/WordPress/gutenberg/pull/61942)).
+-   `label` prop in Actions API can be either a `sting` value or a `function`, in case we want to use information from the selected items. ([#61942](https://github.com/WordPress/gutenberg/pull/61942)).
 
 ## 1.2.0 (2024-05-16)